- When writing metadata to a geom provider, open the it as read-write since it   might do subsequent reads from other providers. This stopped geli (and   probably other classes using g_metadata_store as well) from being put on top   of gvinum raid5 volumes.
Note:
The reason it fails in the gvinum raid5 case is that gvinum will read back the
old parity stripe before calculating the new parity stripe to be written out
again.  The write will then fail because the underlying disk to be read is
opened write only.

MFC after:	1 week