sh: Do not try to execute binary files as scripts.
If execve() returns an [ENOEXEC] error, check if the file is binary before
trying to execute it using sh. A file is considered binary if at least one
of the first 256 bytes is '\0'.

In particular, trying to execute ELF binaries for the wrong architecture now
fails with an "Exec format error" message instead of syntax errors and
potentially strange results.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@218320 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f