Handle the corner case in vm_fault_quick_hold_pages().
If supplied length is zero, and user address is invalid, function
might return -1, due to the truncation and rounding of the address.
The callers interpret the situation as EFAULT. Instead of handling
the zero length in caller, filter it in vm_fault_quick_hold_pages().

Sponsored by:	The FreeBSD Foundation
Reviewed by:	alc


git-svn-id: svn+ssh://svn.freebsd.org/base/head@220001 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f