First off: update the driver README, the old one was horribly crusty, and this still isn't perfect, but its at least a bit more recent.
Secondly, a few improvements to the driver from Andrew Boyer,
support hint to allow devices to not attach, add VLAN_HWTSO
capability so vlans can use TSO, fix in the interrupt handler
to make sure the stack TX queue is processed. Oh, and also
make sure IPv6 does not cause a re-init in the ioctl routine.
Thanks for your efforts Andrew!

Thanks to Claudio Jeker for noticing the ixgbe_xmit() routine
was not correctly swapping the dma map from the first to the
last descriptor in a multi-descriptor transmission, corrected
this.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@222588 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f