Globally replace u_int*_t from (non-contributed) man pages.
The reasoning behind this, is that if we are consistent in our
documentation about the uint*_t stuff, people will be less tempted to
write new code that uses the non-standard types.

I am not going to bump the man page dates, as these changes can be
considered style nits. The meaning of the man pages is unaffected.

MFC after:	1 month


git-svn-id: svn+ssh://svn.freebsd.org/base/head@231564 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f