jilles accurately advises that, in the context of shell redirection, the `:' builtin is processed specially and thus the `: > file' syntax for example will cause premature termination of the current shell on redirection-error.
The `true' builtin on the other-hand is not included in this special
processing (for compatibility reasons to satisfy legacy scripts programmed
for systems where `true' is not a builtin).

Change bare `: > file' syntax into `true > file' syntax to prevent premature
shell termination in the event of redirection-error.

NOTE: Instances of `: > file' that appear within a sub-shell have been left
unmodified as these will not cause premature termination of the main script.

Reviewed by:	jilles, adrian (co-mentor)
Approved by:	adrian (co-mentor)


git-svn-id: svn+ssh://svn.freebsd.org/base/head@240843 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f