Cleanup uio-related code to use struct nvme_request and nvme_ctrlr_submit_io_request().
While here, also fix case where a uio may have more than 1 iovec.
NVMe's definition of SGEs (called PRPs) only allows for the first SGE to
start on a non-page boundary.  The simplest way to handle this is to
construct a temporary uio for each iovec, and submit an NVMe request
for each.

Sponsored by:	Intel


git-svn-id: svn+ssh://svn.freebsd.org/base/head@241661 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f