Changes to resolve races in bpfread() and catchpacket() that, at worst, cause kernel panics.
Add a flag to the bpf descriptor to indicate whether the hold buffer
is in use. In bpfread(), set the "hold buffer in use" flag before
dropping the descriptor lock during the call to bpf_uiomove().
Everywhere else the hold buffer is used or changed, wait while
the hold buffer is in use by bpfread(). Add a KASSERT in bpfread()
after re-acquiring the descriptor lock to assist uncovering any
additional hold buffer races.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@244090 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f