Fix a VAP BSS node reference in the HT code to actually take a reference before using said node.
The "blessed" way here is to take a node reference before referencing
anything inside the node, otherwise the node can be freed between
the time the pointer is copied/dereferenced and the time the node contents
are used.

This mirrors fixes that I've done elsewhere in the net80211/driver
stack.

PR:		kern/178470


git-svn-id: svn+ssh://svn.freebsd.org/base/head@250442 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f