When installing updates, install new directories first and remove old directories last.
This is generally handled by the fact that the list of filesystem objects
is sorted, but this sorting is broken by code which moves .so files ahead
(so that they're present before any binaries which use them)... that code
also moved .so files ahead of directories, which is a problem for upgrading
to 10.0 where there's a new directory containing new .so files.

Errata Notice Candidate.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@256646 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f