Call cpu_setup() immediately after the page tables are installed.  This enables data cache and other chip-specific features.  It was previously done via an early SYSINIT, but it was being done after pmap and vm setup, and those setups need to use mutexes.  On some modern ARM platforms, the ldrex/strex instructions that implement mutexes require the data cache to be enabled.
A nice side effect of enabling caching earlier is that it eliminates the
multi-second pause that used to happen early in boot while physical memory
and pmap and vm were being set up.  On boards with 1 GB or more of ram
this pause was very noticible, sometimes 5-6 seconds.

PR:		arm/183740


git-svn-id: svn+ssh://svn.freebsd.org/base/head@258392 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f