Put the release objdir inside the chroot.
When building the system from outside the chroot as part of the release
process, ensure the objdir is within the chroot so that the whole output
of the release.sh script is self-contained in a single directory.

Use CHROOTDIR/tmp/obj instead of CHROOTDIR/usr/obj to minimize possible
interferences with the output of the build itself.

Reviewed by:	gjb
Approved by:	rpaulo (mentor)


git-svn-id: svn+ssh://svn.freebsd.org/base/head@259881 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f