Check for mismatched vref()/vdrop()
Assert that the hold count has not fallen below the use count, a situation
that would only happen when a vref() (or similar) is erroneously paired
with a vdrop().  This situation has not been observed in the wild, but
could be helpful for someone implementing a new filesystem.

Reviewed by:	kib
Approved by:	hrs (mentor)


git-svn-id: svn+ssh://svn.freebsd.org/base/head@266482 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f