Handle multiple calls to rxq_eof for single packet completion
This requires the VMware vmxnet3 device to flip the start of packet
descriptor's generation before the rest of the packet's descriptors
have been loaded into the Rx ring. I've never observed this behavior,
and it seems to make the most sense not to do it this way. But it is
not a lot of work for the driver to handle this situation just in case.

MFC after:	1 week


git-svn-id: svn+ssh://svn.freebsd.org/base/head@267662 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f