Prefer vt(4) for UEFI boot
The UEFI framebuffer driver vt_efifb requires vt(4), so add a mechanism
for the startup routine to set the preferred console.  This change is
ugly because console init happens very early in the boot, making a
cleaner interface difficult.  This change is intended only to facilitate
the sc(4) / vt(4) transition, and can be reverted once vt(4) is the
default.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@268158 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f